+++ python/lsst/dax/ppdb/bigquery/_manifest.py
